LayoutTests/fast/events/zoom-dblclick-expected.txt



Zoomed
PASS event.clientX is 167
PASS event.clientY is 83
PASS event.pageX is 167
PASS event.pageY is 83
PASS successfullyParsed is true

TEST COMPLETE
